removed which word from chapter1
Abilities are powers an agent has to perform various actions.[1] Some abilities are very common among human agents, like the ability to walk or to speak. Other abilities are only possessed by a few, such as the ability to perform a double backflip or to prove Gödel's incompleteness theorem. While all abilities are powers, the converse is not true, i.e. there are some powers that are not abilities. This is the case, for example, for powers that are not possessed by agents, like the power of salt to dissolve in water. But some powers possessed by agents do not constitute abilities either. For example, the power to understand French is not an ability in this sense since it does not involve an action, in contrast to the ability to speak French.[1] This distinction depends on the difference between actions and non-actions. Actions are usually defined as events that an agent performs for a purpose and that are guided by the person's intention,[2][3] in contrast to mere behavior, like involuntary reflexes.[4][5] In this sense, abilities can be seen as intelligent powers.

Various terms within the semantic field of the term "ability" are sometimes used as synonyms but have slightly different connotations. Dispositions, for example, are often equated with powers and differ from abilities in the sense that they are not necessarily linked to agents and actions.[1][6] Abilities are closely related to know-how, as a form of practical knowledge on how to accomplish something. But it has been argued that these two terms may not be identical since know-how belongs more to the side of knowledge of how to do something and less to the power to actually do it.[7][1] The terms "aptitude" and "talent" usually refer to outstanding inborn abilities.[8] They are often used to express that a certain set of abilities can be acquired when properly used or trained. Abilities acquired through learning are frequently referred to as skills.[9] The term "disability" is usually used for a long-term absence of a general human ability that significantly impairs what activities one can engage in and how one can interact with the world.[10] In this sense, not any lack of an ability constitutes a disability. The more direct antonym of "ability" is "inability" instead.[11]
Various theories ofe essential features of abilities have been proposed. The conditional analysis is the traditionally dominant approach. It defines abilities in terms of what one would do if one had the volition to do so. For modal theories of ability, by contrast, having an ability means that the agent has the possibility to execute the corresponding action. Other approaches include defining abilities in terms of dispositions and potentials. While all the concepts used in these different approaches are closely related, they have slightly different connotations, ich often become relevant for avoiding various counterexamples.